move damage to assets

+++ Assets/Scripts/Enemy/EnemyAsset.cs
@@ -5,7 +5,7 @@ namespace Enemy
     [CreateAssetMenu(menuName = "Assets/Enemy Asset", fileName = "Enemy Asset")]
     public class EnemyAsset : ScriptableObject
     {
-        public int StartHealth;
+        public float StartHealth;
         public EnemyView ViewPrefab;
         public bool isFlyingEnemy;
     }
